Run curl in insecure mode
Without this we get a warning that:

---
curl: (60) SSL certificate problem: unable to get local issuer certificate
More details here: https://curl.haxx.se/docs/sslcerts.html

curl failed to verify the legitimacy of the server and therefore could not
establish a secure connection to it. To learn more about this situation and
how to fix it, please visit the web page mentioned above.
Error: Process completed with exit code 60.
---

It's probably not ideal to just turn off checking, but as all we're
doing is grabbing a list of members to compare with Wikidata, it should
hopefully be fine.